requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-auth
jinja2
python-multipart